libc_support+.cc: derive noexcept specifiers
Wether a function declared in a C header is noexcept depends on the
library that provides the C header and, further on the compiler, as
different compilers treat C functions that are not declared nothrow
differently.

Thus, in order to not have to rely on one or the other, derive the
noexcept attribute for all functions that implement C library
functionality.

Change-Id: If0593a52b8e84aece01c36b89a8828beaacb8bc2
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>