region.cc: consider alignment also for size, not only for start
The methods Region_list::find_free() and Region_list::find_free_rev()
have to consider the alignment also for the size, not only for the start
address.

These functions are exclusively used by Memory::find_free_ram() and
Memory::find_free_ram_rev().

In particular when using the default alignment of L4_PAGESIZE, it is
**essential** to consider the alignment for the size: First, the KIP
memory descriptors can only handle memory regions with a granularity of
1 KiB. Second, Sigma0 must round down the start address of each
conventional memory region with page granularity and round up the end
address of the region with page granularity.

Of course this alignment is only relevant for regions which are written
to the KIP, so for regions not written to the KIP, we might waste a few
bytes -- which shouldn't matter.

Change-Id: If145b7a3ffd6af5869b6c7cfaf9b574c1a74d570
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>